moves service dir to internal
