+++ v1/zenaton/internal/service/http.go
@@ -12,10 +12,12 @@ var client = http.Client{
 	},
 }
 
+// Get sends a GET request to the specified url
 func Get(url string) (*http.Response, error) {
 	return client.Get(url)
 }
 
+// Post sends a json POST http request to the specified url with the specified body
 func Post(url string, body interface{}) (*http.Response, error) {
 	jsonBody, err := json.Marshal(body)
 	if err != nil {
@@ -25,6 +27,7 @@ func Post(url string, body interface{}) (*http.Response, error) {
 	return client.Post(url, "application/json", bytes.NewBuffer(jsonBody))
 }
 
+// Put sends a json PUT http request to the specified url with the specified body
 func Put(url string, body interface{}) (*http.Response, error) {
 	jsonBody, err := json.Marshal(body)
 	if err != nil {
