+++ v1/zenaton/internal/service/service_test.go
@@ -1,4 +1,4 @@
-package service_test
+package service
 
 import (
 	"fmt"
