+++ v1/zenaton/task/taskmanager.go
@@ -2,7 +2,7 @@ package task
 
 import (
 	"fmt"
-	"github.com/zenaton/zenaton-go/v1/zenaton/service/serializer"
+	"github.com/zenaton/zenaton-go/v1/zenaton/internal/service/serializer"
 	"sync"
 )
